+++ app/components/map/layers/MapAircraftList.vue
@@ -219,34 +219,43 @@ function setVisiblePilots() {
         showTracks.value = tracks;
     }
 
-    if (store.config.airports?.length && store.config.onlyAirportsAircraft) {
+    const dashboardMode = !store.config.onlyAirportsAircraft && store.config.dashboardId;
+
+    if (store.config.airports?.length && (store.config.onlyAirportsAircraft || store.config.dashboardId)) {
         const aircraft = new Set<number>();
         const mode = store.config.airportMode ?? 'all';
+        const airports = new Set<string>();
 
-        for (const key in dataStore.airportsList.value) {
-            if (!store.config.airports!.includes(key)) continue;
+        for (const key of store.config.airports) {
+            airports.add(key);
 
             const airportAircraft = dataStore.airportsList.value[key]?.aircraft;
             const ids = mode === 'ground'
                 ? [...(airportAircraft?.groundDep ?? []), ...(airportAircraft?.groundArr ?? [])]
-                : mode === 'all'
-                    ? Object.values(airportAircraft ?? {}).flat()
-                    : airportAircraft?.[mode as MapAircraftKeys] ?? [];
+                : mode === 'airborne'
+                    ? [...(airportAircraft?.departures ?? []), ...(airportAircraft?.arrivals ?? [])]
+                    : mode === 'all'
+                        ? Object.values(airportAircraft ?? {}).flat()
+                        : airportAircraft?.[mode as MapAircraftKeys] ?? [];
 
             for (const cid of ids) {
                 aircraft.add(cid);
             }
         }
 
-        if (aircraft.size) {
-            dataStore.visiblePilots.value = dataStore.visiblePilots.value.filter(x => aircraft.has(x.cid));
+        if (aircraft.size || dashboardMode) {
+            dataStore.visiblePilots.value = dataStore.visiblePilots.value.filter(x => aircraft.has(x.cid) ||
+                (dashboardMode &&
+                    dataStore.vatsim.data.keyedPilots.value[x.cid]?.departure &&
+                    !airports.has(dataStore.vatsim.data.keyedPilots.value[x.cid]?.departure ?? '') &&
+                    !airports.has(dataStore.vatsim.data.keyedPilots.value[x.cid]?.arrival ?? '')));
         }
         else {
             dataStore.visiblePilots.value = [];
         }
     }
 
-    if (store.config.airport && store.config.onlyAirportAircraft) {
+    if (store.config.airport && (store.config.onlyAirportAircraft || store.config.dashboardId)) {
         const airport = dataStore.airportsList.value[store.config.airport];
         if (airport) {
             const coords = [dataStore.vatspy.value?.data.keyAirports.realIcao[store.config.airport].lon, dataStore.vatspy.value?.data.keyAirports.realIcao[store.config.airport].lat];
@@ -262,6 +271,9 @@ function setVisiblePilots() {
                 if (store.config.airportMode && store.config.airportMode !== 'all') {
                     if (store.config.airportMode === 'ground') {
                         return airport.aircraft.groundArr?.includes(x.cid) || airport.aircraft.groundDep?.includes(x.cid);
+                    }
+                    if (store.config.airportMode === 'airborne') {
+                        return airport.aircraft.departures?.includes(x.cid) || airport.aircraft.arrivals?.includes(x.cid);
                     }
                     else {
                         return airport.aircraft[store.config.airportMode as MapAircraftKeys]?.includes(x.cid);