+++ app/components/map/layers/MapSelect.vue
@@ -299,7 +299,7 @@ const select = useTemplateRef('select');
 useClickOutside({
     element: select,
     callback: () => {
-        if (!isMobileOrTablet.value) return;
+        if (!isMobileOrTablet.value || !map.value) return;
         map.value!.getTargetElement().style.cursor = 'grab';
         openedOverlay.value = null;
         selectFeature(false);
@@ -627,6 +627,7 @@ watch(contextMenu, val => {
 function createSelectHandler(type: EventType, select: Select) {
     return async (arg: SelectEvent) => {
         try {
+            if (!map.value) return;
             const selected = (arg.mapBrowserEvent && type !== 'hover') ? arg.selected : select.getFeatures().getArray();
 
             if (hoverAwaiting && selected.length && selected.length === states[type].selectedFeatures.value.length && selected.every(x => states[type].selectedFeatures.value.includes(x))) return;
@@ -642,7 +643,9 @@ function createSelectHandler(type: EventType, select: Select) {
                 if (type === 'hover' || isMobileOrTablet.value) {
                     sleep(100).then(() => {
                         if (!select.getFeatures().getArray().length) {
-                            map.value!.getTargetElement().style.cursor = 'grab';
+                            if (map.value) {
+                                map.value!.getTargetElement().style.cursor = 'grab';
+                            }
                             openedOverlay.value = null;
                             selectFeature(false);
                         }
@@ -686,7 +689,7 @@ function createSelectHandler(type: EventType, select: Select) {
 
                 tookAction = true;
 
-                if (!map.value!.getTargetElement()) return false;
+                if (!map.value?.getTargetElement()) return false;
 
                 if (result === false) {
                     map.value!.getTargetElement().style.cursor = 'grab';
@@ -753,7 +756,9 @@ function createSelectHandler(type: EventType, select: Select) {
 
             if (!tookAction) {
                 if (type === 'hover') {
-                    map.value!.getTargetElement().style.cursor = 'grab';
+                    if (map.value) {
+                        map.value!.getTargetElement().style.cursor = 'grab';
+                    }
                     select.clearSelection();
                 }
                 else if (type === 'click') {